scripts/python/pyproject.toml
[project]
name = "scripts-python"
version = "0.1.0"
description = "python scripts managed by uv"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.2",
  "gunicorn>=23.0.0",
  "httpx>=0.27.2",
  "pexpect>=4.9.0",
  "pydantic>=2.9.2",
  "python-dotenv>=1.0.1",
  "uvicorn>=0.32.0",
]

[tool.uv]
dev-dependencies = [
  "pytest-cov>=5.0.0",
  "pylint-pytest>=1.1.8",
  "pytest>=8.2.0",
  "pytest-mock>=3.14.0",
  "pytest-testinfra>=10.1.1",
]

[tool.mypy]
show_error_context    = true
show_column_numbers   = true
disallow_untyped_defs = true
no_implicit_optional  = true
warn_return_any       = true
check_untyped_defs    = true
